+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixes the general preprocessor syntax highlighting and adds testing
 - Fixes using function/subroutine as parameter in functions/subroutines
   ([#207](https://github.com/krvajal/vscode-fortran-support/issues/207))
+- Fixes labelled conditionals erroneous highlighting
+  ([#204](https://github.com/krvajal/vscode-fortran-support/issues/204))
 
 ### Added